UI work

+++ src/Binacle.Net.Api.UIModule/Models/PackingResult.cs
@@ -6,6 +6,6 @@ internal class PackingResult
 	public Bin Bin { get; set; }
 	public List<PackedItem>? PackedItems { get; set; }
 	public List<PackedItem>? UnpackedItems { get; set; }
-	public double? PackedItemsVolumePercentage { get; set; }
-	public double? PackedBinVolumePercentage { get; set; }
+	public decimal? PackedItemsVolumePercentage { get; set; }
+	public decimal? PackedBinVolumePercentage { get; set; }
 }
